[doc] Clarify interpretation of `verifyClient` (#1994)
Prior to this commit, the documentation for the `verifyClient` option
was somewhat confusing due to its use of the passive voice and an
uncommon interpretation of the term "arguments".

Explain the way `ws` interprets the value of `verifyClient` using an
active voice and more traditional meanings of "parameters" and
"arguments".